Patch 1.4 - Building - PP
Added a navigation bar.
Added an Administration section to the navigation bar, visible only to administrators and moderators.
The Administration section includes two subsections: Users and Items.
In the Users subsection, all registered users are displayed, and roles can be added or removed.
In the Items subsection, all items are listed, and new items can be added or removed.